Create app0.1(incomplete)
To run the website, you would need to install the Flask framework and run the Python script above using a command such as python app.py. This would start a local web server on your computer that you can access in a web browser at http://localhost:5000 to view the website.